[Chore/#341] NotificationId 타입 Int에서 Long으로 변경

+++ app/src/main/java/org/sopt/official/data/model/notification/response/NotificationDetailResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class NotificationDetailResponse(
     @SerialName("notificationId")
-    val notificationId: Int,
+    val notificationId: Long,
     @SerialName("userId")
     val userId: Int,
     @SerialName("title")
