+++ app/src/main/java/org/sopt/official/domain/usecase/notification/GetNotificationDetailUseCase.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class GetNotificationDetailUseCase @Inject constructor(
     private val notificationRepository: NotificationRepository
 ) {
-    suspend operator fun invoke(notificationId: Int): Result<NotificationDetailResponse> {
+    suspend operator fun invoke(notificationId: Long): Result<NotificationDetailResponse> {
         return notificationRepository.getNotificationDetail(notificationId)
     }
 }
