+++ app/src/main/java/org/sopt/official/domain/usecase/notification/UpdateNotificationReadingStateUseCase.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class UpdateNotificationReadingStateUseCase @Inject constructor(
     private val notificationRepository: NotificationRepository
 ) {
-    suspend operator fun invoke(notificationId: Int): Result<NotificationReadingStateResponse> {
+    suspend operator fun invoke(notificationId: Long): Result<NotificationReadingStateResponse> {
         return notificationRepository.updateNotificationReadingState(notificationId)
     }
 }
